==> code/2_costHosp.R <==
source("code/0_parameters.R")

#### Clean hospitalisations data ####

# Read dataset and remove variable names from Stata file where needed
hosp            <- read_dta("data/hospital.dta")
var_label(hosp) <- NULL

# Change coding for those without scabies from 0 to 2 (for table 2)
hosp$pos_scabies[hosp$pos_scabies==0]   <- 2 

hosp <- hosp %>% 
  # Pre-MDA only
  filter(timept==1) %>% 
  
  mutate(sex       = factor(sex, 
                            levels = c(1,2), 
                            labels = c("Male", "Female")),
         ethnicity = factor(ethnicity, 
                            levels = c(1,2), 
                            labels = c("I-Taukei", "Others")),
         timept    = factor(timept, 
                            levels = c(1,2), 
                            labels = c("Pre", "Post")),
         residence = 0,
         residence = factor(residence, 
                            levels = c(1,2), 
                            labels = c("Urban", "Rural")),
         age_5yr   = cut(enrolage_yrs, 
                         breaks = c(0, 5, 10, 15, 25, 35, 45, 55, 65, Inf),
                         labels = c("0-4", "5-9", "10-14",
                                    "15-24", "25-34", "35-44",
                                    "45-54", "55-64", "65+")),
         age_grp   = fct_collapse(age_5yr, 
                                  "5-14" = c("5-9", "10-14"),  
                                  "15+"  = c("15-24","25-34","35-44",
                                             "45-54","55-64","65+")),
         ageD      = fct_collapse(age_5yr, 
                                  "15+" = c("15-24","25-34","35-44",
                                            "45-54","55-64","65+")),
         txWardAdmit = 1, # everyone admitted to hospital
         txICUAdmit = if_else(!is.na(icudays), true=1, false=0),
         txDiagnostics = if_else(rowSums(pick(c(bc,tissueculture_yn,skin_swabyn)), 
                                         na.rm=T)>=1,
                                 true=1, 
                                 false=0),
         txMedsperOral = if_else(rowSums(pick(c(oralab_checkad___flu, oralab_checkad___pen, 
                                                oralab_checkad___amo, oralab_checkad___met, 
                                                oralab_checkad___ery, oralab_checkad___dox, 
                                                oralab_checkad___cep, oralab_checkad___sep, 
                                                oralab_checkad___oth, oralab_typdc___flu, 
                                                oralab_typdc___pen, oralab_typdc___amo, 
                                                oralab_typdc___met, oralab_typdc___ery, 
                                                oralab_typdc___dox, oralab_typdc___cep, 
                                                oralab_typdc___sep, oralab_typdc___oth)), 
                                         na.rm = T)>=1,
                                 true=1, false=0),
         txMedsInjectn    = if_else(rowSums(pick(c(ivab_check___clox, ivab_check___gent, 
                                                   ivab_check___pen, ivab_check___met, 
                                                   ivab_check___ery, ivab_check___cef, 
                                                   ivab_check___cip, ivab_check___mer, 
                                                   ivab_check___sep, ivab_check___oth)), 
                                            na.rm = T)>=1,
                                    true=1, 
                                    false=0)) %>% 
  
  select(record_id           = record_id,
         sex                 = sex,
         age_integer         = enrolage_yrs,
         age_grp             = age_grp,
         ageD                = ageD,
         ethnicity           = ethnicity,
         residence           = residence,
         timept              = timept,
         admOutcome          = adm_outcome,
         txWardAdmit         = txWardAdmit,
         txICUAdmit          = txICUAdmit,
         txDiagnostics       = txDiagnostics,
         txMedsperOral       = txMedsperOral,
         txMedsInjectn       = txMedsInjectn,
         dHospStay           = los,
         dICUAdmit           = icudays,
         dVentilation        = vent_days,
         dxScabies           = pos_scabies,
         dxImpetigo          = impetigoyn,
         dxCellulitis        = sub_cel,
         dxAbscess           = sub_abs,
         dxNecrotisingFas    = sub_nf,
         dxWoundInf          = sub_wi,
         dxWoundInfSurg      = sub_swi,
         dxImpetigoSevere    = sub_si,
         dxScabiesInf        = sub_if,
         dxScabiesCrus       = sub_cs,
         dxLymphadenitis     = sub_la,
         dxPyomyositis       = sub_my,
         txProcAmputat       = amputation,
         txProcSurgery       = surgery,
         labCultureBlood     = bc,
         labCultureTissue    = tissueculture_yn,
         labSkinSwab         = skin_swabyn,
         ivab_check___clox, ivab_check___gent, ivab_check___pen, ivab_check___met, 
         ivab_check___ery, ivab_check___cef, ivab_check___cip, ivab_check___mer, 
         ivab_check___sep, ivab_check___oth, oralab_checkad___flu, oralab_checkad___pen, 
         oralab_checkad___amo, oralab_checkad___met, oralab_checkad___ery,
         oralab_checkad___dox, oralab_checkad___cep, oralab_checkad___sep, 
         oralab_checkad___oth, oralab_typdc___flu, oralab_typdc___pen, 
         oralab_typdc___amo, oralab_typdc___met, oralab_typdc___ery, 
         oralab_typdc___dox, oralab_typdc___cep, oralab_typdc___sep, 
         oralab_typdc___oth, oralabduration, oralabdc_days,
         dInjnMeds            = ivabduration) 

# Change missing values to 0 in dummy variables
hosp$dICUAdmit[is.na(hosp$dICUAdmit)]           <- 0
hosp$dVentilation[is.na(hosp$dVentilation)]     <- 0
hosp$oralabduration[is.na(hosp$oralabduration)] <- 0
hosp$dInjnMeds[is.na(hosp$dInjnMeds)]           <- 0

# Change empty cells to 0, text cells to number
hosp$oralabdc_days[hosp$oralabdc_days==""] <- "0"
hosp$oralabdc_days <- gsub("[^0-9]", "", hosp$oralabdc_days)

# Factor admission outcomes and calculate total number of medications
hosp <- hosp %>% 
  mutate(admOutcome = factor(x = admOutcome, levels = c(1,2,3,4), 
                             labels = c("Discharged", "Transferred", 
                                        "Disability", "Death"))) %>% 
  mutate(across(c(dxImpetigo:dInjnMeds), as.numeric)) %>% 
  mutate(dOralMeds   = rowSums(across(c(oralabduration, oralabdc_days)), 
                               na.rm=TRUE),
         numInjnMeds = rowSums(across(c(ivab_check___clox, ivab_check___gent, 
                                        ivab_check___pen, ivab_check___met, 
                                        ivab_check___ery, ivab_check___cef, 
                                        ivab_check___cip, ivab_check___mer, 
                                        ivab_check___sep, ivab_check___oth)),
                               na.rm=TRUE),
         numOralMeds = rowSums(across(c(oralab_checkad___flu, oralab_checkad___pen, 
                                        oralab_checkad___amo, oralab_checkad___met, 
                                        oralab_checkad___ery, oralab_checkad___dox, 
                                        oralab_checkad___cep, oralab_checkad___sep, 
                                        oralab_checkad___oth, oralab_typdc___flu, 
                                        oralab_typdc___pen, oralab_typdc___amo, 
                                        oralab_typdc___met, oralab_typdc___ery, 
                                        oralab_typdc___dox, oralab_typdc___cep, 
                                        oralab_typdc___sep, oralab_typdc___oth)),
                               na.rm=TRUE))   


hosp <- hosp %>%
  # DaysOfParticularMed = I / TotalNumberOfMeds * TotalDaysOfMeds
  # I is 1 if patient was given particular med, 0 otherwise
  mutate(dInjnCloxac = if_else(numInjnMeds!=0, 
                               true=ivab_check___clox/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnGentam = if_else(numInjnMeds!=0, 
                               true=ivab_check___gent/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnPenici = if_else(numInjnMeds!=0, 
                               true=ivab_check___pen/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnMetron = if_else(numInjnMeds!=0, 
                               true=ivab_check___met/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnErythr = if_else(numInjnMeds!=0, 
                               true=ivab_check___ery/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnCeftri = if_else(numInjnMeds!=0, 
                               true=ivab_check___cef/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnCiprof = if_else(numInjnMeds!=0, 
                               true=ivab_check___cip/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnMerope = if_else(numInjnMeds!=0, 
                               true=ivab_check___mer/numInjnMeds*dInjnMeds, 
                               false=0),
         dInjnOthers = if_else(numInjnMeds!=0, 
                               true=ivab_check___oth/numInjnMeds*dInjnMeds, 
                               false=0),
         dOralFluclo = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___flu+oralab_typdc___flu)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralPenici = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___pen+oralab_typdc___pen)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralAmoxyc = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___amo+oralab_typdc___amo)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralMetron = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___met+oralab_typdc___met)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralErythr = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___ery+oralab_typdc___ery)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralDoxycy = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___dox+oralab_typdc___dox)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralCephal = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___cep+oralab_typdc___cep)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralSeptri = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___sep+oralab_typdc___sep)/
                                 numOralMeds*dOralMeds, 
                               false=0),
         dOralOthers = if_else(numOralMeds!=0, 
                               true=(oralab_checkad___oth+oralab_typdc___oth)/
                                 numOralMeds*dOralMeds, 
                               false=0)) %>% 
  # Different quantities of medications required for different age groups
  mutate(qInjnCloxac = if_else(ageD=="0-4", 
                               true=medList$qInjnCloxac_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnCloxac_05.09,
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnCloxac_10.14, 
                                               false=medList$qInjnCloxac_15plus))),
         qInjnGentam = if_else(ageD=="0-4", 
                               true=medList$qInjnGentam_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnGentam_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnGentam_10.14, 
                                               false=medList$qInjnGentam_15plus))),
         qInjnPenici = if_else(ageD=="0-4", 
                               true=medList$qInjnPenici_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnPenici_05.09, 
                                       if_else(ageD=="10-14", 
                                               true =medList$qInjnPenici_10.14, 
                                               false=medList$qInjnPenici_15plus))),
         qInjnMetron = if_else(ageD=="0-4", 
                               true=medList$qInjnMetron_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnMetron_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnMetron_10.14, 
                                               false=medList$qInjnMetron_15plus))),
         qInjnErythr = if_else(ageD=="0-4", 
                               true=medList$qInjnErythr_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnErythr_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnErythr_10.14, 
                                               false=medList$qInjnErythr_15plus))),
         qInjnCeftri = if_else(ageD=="0-4", 
                               true=medList$qInjnCeftri_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnCeftri_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnCeftri_10.14, 
                                               false=medList$qInjnCeftri_15plus))),
         qInjnCiprof = if_else(ageD=="0-4", 
                               true=medList$qInjnCiprof_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnCiprof_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnCiprof_10.14, 
                                               false=medList$qInjnCiprof_15plus))),
         qInjnMerope = if_else(ageD=="0-4", 
                               true=medList$qInjnMerope_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnMerope_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnMerope_10.14, 
                                               false=medList$qInjnMerope_15plus))),
         qInjnOthers = if_else(ageD=="0-4", 
                               true=medList$qInjnOthers_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qInjnOthers_05.09,
                                       if_else(ageD=="10-14", 
                                               true=medList$qInjnOthers_10.14, 
                                               false=medList$qInjnOthers_15plus))),
         qOralFluclo = if_else(ageD=="0-4", 
                               true=medList$qSuspFlucl_00.04,  
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspFlucl_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralFlucl_10.14, 
                                               false=medList$qOralFlucl_15plus))),
         qOralPenici = if_else(ageD=="0-4", 
                               true=medList$qSuspPenic_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspPenic_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralPenic_10.14, 
                                               false=medList$qOralPenic_15plus))),
         qOralAmoxyc = if_else(ageD=="0-4", 
                               true=medList$qSuspAmoxy_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qOralAmoxy_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralAmoxy_10.14, 
                                               false=medList$qOralAmoxy_15plus))),
         qOralMetron = if_else(ageD=="0-4", 
                               true=medList$qSuspMetro_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspMetro_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralMetro_10.14, 
                                               false=medList$qOralMetro_15plus))),
         qOralErythr = if_else(ageD=="0-4", true=medList$qSuspEryth_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspEryth_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralEryth_10.14, 
                                               false=medList$qOralEryth_15plus))),
         qOralDoxycy = if_else(ageD=="0-4", 
                               true=medList$qOralDoxyc_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qOralDoxyc_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralDoxyc_10.14, 
                                               false=medList$qOralDoxyc_15plus))),
         qOralCephal = if_else(ageD=="0-4", 
                               true=medList$qSuspCepha_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspCepha_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralCepha_10.14, 
                                               false=medList$qOralCepha_15plus))),
         qOralSeptri = if_else(ageD=="0-4", 
                               true=medList$qSuspSeptr_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspSeptr_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralSeptr_10.14, 
                                               false=medList$qOralSeptr_15plus))),
         qOralOthers = if_else(ageD=="0-4", 
                               true=medList$qSuspOthers_00.04, 
                               if_else(ageD=="5-9", 
                                       true=medList$qSuspOthers_05.09, 
                                       if_else(ageD=="10-14", 
                                               true=medList$qOralOthers_10.14, 
                                               false=medList$qOralOthers_15plus)))) %>% 
  
  # Total quantity of meds = Quantity of meds * Duration of treatment
  mutate(qInjnCloxac = ceiling(qInjnCloxac * dInjnCloxac),
         qInjnGentam = ceiling(qInjnGentam * dInjnGentam),
         qInjnPenici = ceiling(qInjnPenici * dInjnPenici),
         qInjnMetron = ceiling(qInjnMetron * dInjnMetron),
         qInjnErythr = ceiling(qInjnErythr * dInjnErythr),
         qInjnCeftri = ceiling(qInjnCeftri * dInjnCeftri),
         qInjnCiprof = ceiling(qInjnCiprof * dInjnCiprof),
         qInjnMerope = ceiling(qInjnMerope * dInjnMerope),
         qInjnOthers = ceiling(qInjnOthers * dInjnOthers),
         qOralFluclo = ceiling(qOralFluclo * dOralFluclo),
         qOralPenici = ceiling(qOralPenici * dOralPenici),
         qOralAmoxyc = ceiling(qOralAmoxyc * dOralAmoxyc),
         qOralMetron = ceiling(qOralMetron * dOralMetron),
         qOralErythr = ceiling(qOralErythr * dOralErythr),
         qOralDoxycy = ceiling(qOralDoxycy * dOralDoxycy),
         qOralCephal = ceiling(qOralCephal * dOralCephal),
         qOralSeptri = ceiling(qOralSeptri * dOralSeptri),
         qOralOthers = ceiling(qOralOthers * dOralOthers),
         qInjnTotals = rowSums(across(c(qInjnCloxac:qInjnOthers))),
         qOralTotals = rowSums(across(c(qOralFluclo:qOralOthers)))) %>%
  # Remove unwanted variables
  select(-c(dInjnCloxac:dOralOthers, ivab_check___clox:oralabdc_days))


#### Cost of hospitalisations #### 

hosp <- hosp %>%
  # Costs of healthcare services
  mutate(txClinicVisit  = 0,
         txMedsTopical  = 0,
         dWardAdmit     = dHospStay  - dICUAdmit,     # days for ward
         cWardAdmit     = dWardAdmit * cWard["base"], # cost for ward
         cICUAdmit      = dICUAdmit  * cICU["base"],  # cost for ICU
         cHospStay      = cWardAdmit + cICUAdmit,     # cost for both
         cClinicVisit   = txClinicVisit * cVisit["base"], # no clinic visit
         cMedsTopical   = txMedsTopical * cPermetCr,  # no topical meds
         # Costs of laboratory tests
         nDiagnostics  = rowSums(pick(labCultureBlood, labCultureTissue, labSkinSwab), 
                                 na.rm = T),
         cCultureBlood  = labCultureBlood   * cLabtest["base"],
         cCultureTissue = labCultureTissue  * cLabtest["base"],
         cSkinSwab      = labSkinSwab       * cLabtest["base"],
         cDiagnostics   = rowSums(pick(cCultureBlood, cCultureTissue, cSkinSwab), 
                                  na.rm = T),
         # Costs of IV antibiotics
         cInjnCloxac = qInjnCloxac * cCloxacIV["base"],
         cInjnGentam = qInjnGentam * cGentamIV["base"],
         cInjnPenici = qInjnPenici * cPeniciIV["base"],
         cInjnMetron = qInjnMetron * cMetronIV["base"],
         cInjnErythr = qInjnErythr * cErythrIV["base"],
         cInjnCeftri = qInjnCeftri * cCeftriIV["base"],
         cInjnCiprof = qInjnCiprof * cCiprofIV["base"],
         cInjnMerope = qInjnMerope * cMeropeIV["base"],
         cInjnOthers = qInjnOthers * cOthersIV["base"],
         # Costs of oral antibiotics (only 0-4 take syrup amoxil)
         cOralFluclo = if_else(ageD=="15+", 
                               true=qOralFluclo*cFlucloTb["base"], 
                               false=qOralFluclo*cFlucloSn["base"]),
         cOralPenici = if_else(ageD=="15+", 
                               true=qOralPenici*cPeniciTb["base"], 
                               false=qOralPenici*cPeniciSn["base"]),
         cOralAmoxyc = if_else(ageD=="0-4", 
                               true=qOralAmoxyc*cAmoxycSn["base"], 
                               false=qOralAmoxyc*cAmoxycTb["base"]),
         cOralMetron = if_else(ageD=="15+", 
                               true=qOralMetron*cMetronTb["base"], 
                               false=qOralMetron*cMetronSn["base"]),
         cOralErythr = if_else(ageD=="15+", 
                               true=qOralErythr*cErythrTb["base"], 
                               false=qOralErythr*cErythrSn["base"]),
         cOralDoxycy = qOralDoxycy*cDoxycyTb["base"], # not for children
         cOralCephal = if_else(ageD=="15+", 
                               true=qOralCephal*cCephalTb["base"], 
                               false=qOralCephal*cCephalSn["base"]),
         cOralSeptri = if_else(ageD=="15+", 
                               true=qOralSeptri*cSeptriTb["base"], 
                               false=qOralSeptri*cSeptriSn["base"]),
         cOralOthers = if_else(ageD=="15+", 
                               true=qOralOthers*cOthersTb["base"],
                               false=qOralOthers*cOthersSn["base"]),
         # Total costs of all medications
         cMedsInjectn = rowSums(across(c(cInjnCloxac:cInjnOthers))),
         cMedsperOral = rowSums(across(c(cOralFluclo:cOralOthers))),
         cTotalMeds  = cMedsTopical + cMedsInjectn + cMedsperOral,
         cMeanTotal  = rowSums(pick(cHospStay, cDiagnostics, cTotalMeds), na.rm = T)) 




#### Sensitivity analysis ####

# Low and High unit cost of ward bed day
hosp <- hosp %>%
  mutate(cWardAdmit_low    = dWardAdmit    * cWard["low"],
         cWardAdmit_high   = dWardAdmit    * cWard["high"],
         cICUAdmit_low     = dICUAdmit     * cICU["low"],
         cICUAdmit_high    = dICUAdmit     * cICU["high"],
         cDiagnostics_low  = nDiagnostics * cLabtest["low"],
         cDiagnostics_high = nDiagnostics * cLabtest["high"],
         cMeanTotal_cWardAdmit_low    = rowSums(pick(cWardAdmit_low, cICUAdmit, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cWardAdmit_high   = rowSums(pick(cWardAdmit_high, cICUAdmit, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cICUAdmit_low     = rowSums(pick(cWardAdmit, cICUAdmit_low, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cICUAdmit_high    = rowSums(pick(cWardAdmit, cICUAdmit_high, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cDiagnostics_low  = rowSums(pick(cWardAdmit, cICUAdmit, 
                                                     cDiagnostics_low, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cDiagnostics_high = rowSums(pick(cWardAdmit, cICUAdmit, 
                                                     cDiagnostics_high, cTotalMeds), 
                                                na.rm = T))

# Low and High unit cost of ward bed day
hosp <- hosp %>%
  mutate(cWardAdmit_low    = dWardAdmit    * cWard["low"],
         cWardAdmit_high   = dWardAdmit    * cWard["high"],
         cICUAdmit_low     = dICUAdmit     * cICU["low"],
         cICUAdmit_high    = dICUAdmit     * cICU["high"],
         cDiagnostics_low  = nDiagnostics * cLabtest["low"],
         cDiagnostics_high = nDiagnostics * cLabtest["high"],
         cMeanTotal_cWardAdmit_low    = rowSums(pick(cWardAdmit_low, cICUAdmit, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cWardAdmit_high   = rowSums(pick(cWardAdmit_high, cICUAdmit, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cICUAdmit_low     = rowSums(pick(cWardAdmit, cICUAdmit_low, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cICUAdmit_high    = rowSums(pick(cWardAdmit, cICUAdmit_high, 
                                                     cDiagnostics, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cDiagnostics_low  = rowSums(pick(cWardAdmit, cICUAdmit, 
                                                     cDiagnostics_low, cTotalMeds), 
                                                na.rm = T),
         cMeanTotal_cDiagnostics_high = rowSums(pick(cWardAdmit, cICUAdmit, 
                                                     cDiagnostics_high, cTotalMeds), 
                                                na.rm = T))


save(hosp, file = "data/hosp.Rdata")
